Switch to using WAL for ironic sqlite database
The default mode does not work well when several threads and/or processes
are reading and writing to the same database. This change switches
to the WAL mode available since sqlite 3.7.0.

The mode is persistent for a file until changed, so we pre-create
the database file when building a container.