Add self configuration to Ironic for Openshift Baremetal Operator.
We're planning to run these containers along side the BMO deployment.
This means we need to configure Ironic at startup, along with any
images needed etc.  We still need the dualboot image in here but
we are much more limited running in a kubernetes pod than with podman
on the local host.