[WIP] use openshift content
Building the image using openshift forked ironic and ironic-lib
repos.

This is an experimental patch and it should not be merged into
master as is, but rather go into a different branch.

The Dockerfile contains mechanism to compile RPMs from content
coming from the forked ironic and ironic-lib repos to test
image building and running.

Using Python 2 as centos7 doesn't have all the packages required
to run Ironic with Python 3.