Add set -e to all the entrypoint scripts
I noticed that it's possible for the dbsync in the conductor script
to fail (if mariadb is not yet up), which is silently ignored and
results in a broken Ironic.

It's probably one of many possible failures, so lets set -e so containers
exit when an error happens, which allows for automated restart e.g
via k8s or systemd.