[WIP] Yet another build from source
This patch builds ironic and its dependencies (and the
dependencies of the dependencies and... well you get it)
from source.
This is done to reduce the usage of rpms, unless strictly
necessary and only if the rpms are available in standard
base repos, so to simulate a possible process where
external repos are not accessible.
We download all the source packages in a common dir and
build and install them from there.
The build process works as-is, but some deps are probably
missing, so a run test is necessary to determine what we
need to add in terms of python libs from rpms.
In a common internal close building infrastructure it
might happen that the pypy repo is not available but
there are ther ways to get the source packages
in that case, I'm not entirely crazy :)