Retry dbsync until success
I've seen some situations where mariadb is not up soon enough for the
dbsync, despite the 10 retries done inside ironic-dbsync.

In this situation we start the ironic services anyway and since #80 got reverted,
we can't rely on pod/container restart to handle this, so lets retry inside
the script logging a warning instead.