Adds initial redfish UEFI virtual media boot
In order to boot instances via virtual media in UEFI,
we need to be able to master an EFI image which contains
the bootloader and related files.

In this case, we're collecting the shim and grubx64.efi
into the image which will be used by the EFI firmware to
load the bootloader.