Temporarily disable cleaning as a workaround
The HPE virtual install device presents a virtual, emulated, read-only
spinning disk which fails cleaning operaitons in preparation for deployment.

This device is bios controlled, and is unable to be disabled by the user.
The bios may hide it in some cases if previously disabled before the bios
version is updated, or if "intellegent provisioning" has completed on the
host as the volume is intended for installation of other operating
systems.

Some more information can be found in storyboard:

https://storyboard.openstack.org/#!/story/2007229

This patch will likely be reverted once a fix for the device handling is
available in ironic-python-agent for downstream consumption.

rhbz: #1796141