Support explicit allowlist for DHCP clients
In OpenShift, we have a pivot from a "bootstrap" VM to the cluster
Metal3 pod, but there's a short time where both could exist. The
bootstrap knows which mac's it needs to respond to, which is only the
k8s control plane hosts. This adds an optional feature to our dnsmasq
container that blocks DHCP requests from anything but hosts in the
allowlist.

We are not using dnsmasq native features for ignoring clients as DHCPv6
clients is done based on DUID, so the workaround is to block using
iptables.