Use snponly.efi instead of ipxe.efi for IPv4
As part of the UEFI stack, a network stack is available. The iPXE
binary that implements this stack, and only this stack support is
snponly.efi, where as ipxe.efi instead tries to load network driers
that have been authored and incorporated into the ipxe.efi binary
itself.

These drivers are redundant in UEFI mode, and we should not need
them. As such, we should realistically only be using snponly.efi.

If not used, we get an error about no networks having been found
upon loading ipxe.efi.

rhbz# 1830161