Disable sensor data collection
Sensor data collection was intended for an earlier use of
the ironic-image and resulting ironic container image in order
to supply data to prometheus about hardware that has not yet
been deployed upon. Since the deicsion was made not to leverage
the data, it seems pointless to collect sensor data.

As such, disable sensor data collection.