Reload httpd on certificate change
When the certificates used by httpd change (e.g., due to automatic
certificate renewal by cert-manager). We want to reload httpd so it uses
the new certificates. This is achieved by using inotify to get informed
about changes of the files.

As the secrets in k8s are just symlinks to the real files, the event we
see when the certificates are renewed is DELETE_SELF.