Update http_basic configuration
Let's use a variable that will define the auth_strategy for ironic
`IRONIC_AUTH_STRATEGY` that is default to noauth.

The variables for username and password are default to "", if you
are using `http_basic` they need to be set (will fail if one of the
variables is not set).

Update the json_rpc configuration, we will use username/password
instead of http_basic_username and http_basic_password, see [1]

[1] https://review.opendev.org/#/c/742474/