Configure ironic-conductor to auth to ironic-inspector
The ironic-conductor process needs to authenticate to the
ironic-inspector API, so we need to pass the auth config file options to
it also.

This is somewhat complicated by the fact that by default, ironic-api
picks up both the ironic-dist.conf and ironic.conf files, so to maintain
behaviour when passing a list of config files explicitly we need to pass
both. (I have no idea where or how the default is configured.)

However, ironic-conductor does *not* use ironic-dist.conf by default,
and including it breaks CI. Therefore we pass ironic-dist.conf
explicitly to only ironic-api, and don't include it in the list that we
pass to both ironic-api and ironic-conductor.