Explicitly set json_rpc.auth_strategy to noauth
Using crudini to merge this into ironic.conf no longer
works as it gets ignored as it matches the DEFAULT and
it is redundant. But later we then change the DEFAULT in
set_http_basic_server_auth_strategy so we need to set
json_rpc.auth_strategy back to noauth.