Switch ironic configuration rendering to Jinja2
This centralizes the cryptic configuration done in multiple files.
In addition, it configures json_rpc to be authenticated when
API and conductor are running in the same container and
HTTP_BASIC_HTPASSWD is set. API and json_rpc would accept the
same credentials. This is to prevent other processes in the same
host to access the RPC interface over localhost.