Use [] around IPv6 address for conductor host config
Work around https://storyboard.openstack.org/#!/story/2008288 by setting
the DEFAULT.host config option to the host portion of a URL, meaning
that IPv6 addresses are surrounded by [], rather than just the raw IP
address. This is how it is ultimately used inside Ironic.

This solution works both with and without the fix for the issue in
Ironic (https://review.opendev.org/759877).

Fixes #210