fix: ensure dualboot.ipxe readable for ipxe
In order for ipxe to read this dualboot.ipxe, the file has to be readable.
This can be a problem if the image was built on systems with a strict umask
such as 0027. Ensure the file is readable using chmod.

Otherwise one might see errors like the following:

```
Filename: http://21.0.12.28:6180/dualboot.ipxe
http://21.0.12.28:6180/dualboot.ipxe... Operation not permitted (http://ipxe.org/410c618f)
```