Merge ironic and ironic-inspector repositories
This change moves Dockerfile, configuration and scripts from the
ironic-inspector container image repository to the ironic one.

This approach provides two separate Dockerfiles in the same repository,
one for ironic and one for ironic-inspector, that share part of the
logic needed to build the images.
The scripts and configurations are kept separated for the time being as
some of them conflict with each other (e.g. runhttpd).

The original Dockerfile is kept until we make the necessary changes to
the CI.