Merge ironic and ironic-inspector repositories
This change moves Dockerfile, configuration and scripts from the
ironic-inspector container image repository to the ironic one.

This approach provides one Dockerfile that builds an image that contains
all the services necessary to run both ironic and ironic-inspector.
The scripts and configurations are kept separated for the time being.

Those changes will need to be reflected in CI as well.