Don't always pass ignition config in PXE kernel args
When we get images from the image builder, they will contain the
Ignition file. Only pass an Ignition URL on the kernel command line if
the configure-coreos-ipa script has been run to create an ignition file
locally. We will not run this script when using the image builder.

However, we always want to set the rootfs location in the kernel args,
even when we haven't run configure-coreos-ipa because we are using the
image builder instead.