Add a new entry point for the Ironic proxy
In OpenShift we deploy Ironic only on one of the controlplane nodes.
When the provisioning network is disabled, IPA talks to Ironic API via
the node's IP address. If the Metal3 pod is relocated, IPA loses contact
to Ironic API.

Upstream in Metal3, this problem is solved via keepalived. We don't want
to introduce another VIP in OpenShift, so instead this change adds
a new script that launches httpd as a proxy to real Ironic. It will be
deployed as a DaemonSet, making Ironic API available on the API VIP.

The cluster-baremetal-operator will be responsible for updating
the DaemonSet with the up-to-date real Ironic IP address.