Allow customizing public ports for Ironic and Inspector
While 6385 and 5050 are reasonably established, an operator may choose
to assign different ports (e.g. in case of a conflict).

This change also makes it possible to have different listen and access
ports for the rare case when there is another indirection between
consumers and API (e.g. in case of OpenShift we want to have another
reverse proxy on control plane nodes).