Removed ServerName from VirtualHost Directives
ServerName does not support IPv6 addresses, the verification
fails when the second segment doesn't begin in a digit(it doesn't
look enough like a port number).

We only have a single VirtualHost per "Listen" so it is the default.
Remove the use of ServerName as we're not using it to differentiate
between VirtualHosts.