Don't log OS_ prefixed variables to a file
This change stops saving the OS_ prefixed variables to a file but keeps
them logged on stdout. The file is not consumed elsewhere, by either
this image or Ironic itself. Outputting to a container instance's
stdout allows some other log collector to save what may be helpful
debugging information about the Oslo configuration, and adding the:
`|| true` allows for #370's suggestion of `set -euxo pipefail` to work
in cases where the environment variables are not set and grep would
return non-zero.