ironic.conf.j2: Bump min_command_interval to 30 on SCOS
Intended as a temporary change, this bumps the min_command_interval
to 30 on SCOS.
The infrastructure that the OKD Streams build system is going to be
deployed to uses an IPMI proxy which introduces additional latency on
that specific infrastructure.
Increase min_command_interval until a faster proxy is used.
Only affects OKD/SCOS builds of ironic-image.