Add missing authentication to runironic
If Ironic is not configured with TLS and it is also not configured to
have API and Conductor split, access to the API is not protected by any
authentication as it is exposed directly. Ironic API is also listening
in host network. In case the node is not behind a firewall, the API
could be accessed by anyone via network.

The change copies the logic from runinspector to ensure that the
all-in-one script also uses authentication.