python-ironicclient debug container image
The Metal³ ironic container image does not include an ironic-client
installation. Furthermore there is no longer a package in centos
that provides the ironic-client via dnf.

This commit adds a Dockerfile for a docker container image which
contains an installation of ironic-pythonclient. The docker image
is configured such that it can be used as debug container alongside
the Metal³ ironic container, providing an interactive shell
environment which can be used to interact with Metal³ ironic using
the "baremetal" cli.

Signed-off-by: Jan Klippel <jan.klippel@uhurutec.com>