Download and install sources from local dir
Since in CI we use the --no-index option to emulate a pip
disconnected environment to try and reproduce the downstream
build conditions, it's not possible to test normal dependencies
from source as pip won't be able to retrieve them from any
remote source.
To work around that we download all the packages first removing
the --no-index option but using the --no-deps option, preventing
downloading dependencies.
This forces to download only the packages specified in the
requirements file, ensuring total control on main libraries and
dependencies and allowing us to be as granular as needed, easily
switching between RPMs and source packages for testing or even
for downstream builds.
When we install them afterwards if any dependency is missing
the installation process will fail in CI, allowing us to correct
the dependencies list directly the change PR.