Fix runlogwatch script
Use correct values for read command, and remove colors from
output.
a pyinotify event looks like this:
<Event dir=False mask=0x8 maskname=IN_CLOSE_WRITE name=mylogs.gzip path=/shared/log/ironic/deploy pathname=/shared/log/ironic/deploy/mylogs.gzip wd=1 >
by default the output contains colorize text that is
interpreted literally by variable conversion in the
tar command, so not only we have to parse it, but
we also need to get just raw format output.

Signed-off-by: Riccardo Pittau <elfosardo@gmail.com>