remove faulty NUMPROC logic
This removes the faulty NUMPROC processor count detection logic from
the configure-ironic.sh. It used to limit CPU count to 4, but the
"fixed" version made it a MINIMUM of 4.

This detection has now been merged on the ironic api side, so we can
remote the note + the faulty logic, and just leave the minimal env
var there. 0 means ironic will run its autodetection logic and set
maximum of 4 cores.

This fix has been merged on the Ironic side already in 2020, so we
can backport this to all branches we have.

Signed-off-by: Tuomo Tanskanen <tuomo.tanskanen@est.tech>