do not bump major version of github actions
Change dependabot config not to try major bumps of github actions.
We now have grouped bumps to lessen the noise from dependabot, and
if in that group, there is a major bump (like right now a
golangci-lint v2 major bump in many repos) which isn't compatible as is
and requires manual changes, it blocks bumping all the other actions as
well.

Also, change the name of the all-github-actions group, as how it is
inserted in PR title looks wrong. Plain "github-actions" group name
works better there.

Signed-off-by: Tuomo Tanskanen <tuomo.tanskanen@est.tech>