Introduce get_interface_of_ip to look PROVISIONING_IP up
Add a new get_interface_of_ip function that returns the name of the
interface of a given IP, and use it to determine if the
PROVISIONING_IP address is actually present on a network interface.
This improves the code readability and enables additional debugging
output.

This function supersedes a similar check present in
get_provisioning_interface, hence this commit also reverts commit
2742439.

Signed-off-by: Marco Chiappero <marco.chiappero@suse.com>