Update ironic.conf to leverage IRONIC_IPV6
The ironic.conf file allows to control how the ironic process should bind
and communicate, in particular through the "my_ip" and "host_ip" settings.

Now that IRONIC_IPV6 is available, the first one can be augmented setting
"my_ipv6" as well, as per the Ironic documentation:

  "This field [my_ip] does accept an IPv6 address as an override for
   templates and URLs, however it is recommended that [DEFAULT]my_ipv6
   is used along with DNS names for service URLs for dual-stack
   environments."

Also set host_ip to an IPv6 address when found and prioritize it over IPv6
when LISTEN_ALL_INTERFACES is not set to true.

Since the value of host_ip is determined twice within the ironic.conf.j2
template file, by means of long and hard to read set of conditions, avoid
this duplication and improve readability by exporting the correct value
once in scripts/configure-ironic.sh.

Signed-off-by: Marco Chiappero <marco.chiappero@suse.com>