Refactor and improve input validation in wait_for_interface_or_ip
How Ironic and related services are configured and bound to the system
interfaces fundamentally depends on three possible inputs:
- a provisioning IP (if PROVISIONING_IP is set), which, if found, is used to
  determine IRONIC_IP
- a provisioning interface (if PROVISIONING_INTERFACE or
  PROVISIONING_MACS is set), by setting IRONIC_IP to the first IP
  address found on such interface
- IRONIC_IP directly, whenever set

Depending on the variable(s) set, the function might search and possibly wait
for the requested interface or IP to appear.

Rework the test logic of the function, so that the three possible cases
are clearly listed and tested separately, by introducing an explicit
check for IRONIC_IP. This allows to:
- validate the value of IRONIC_IP, when externally provided
- test PROVISIONING_INTERFACE separately, in order to improve readability
- validate at least one of the above three inputs is provided and fail
  gracefully otherwise

Note that 1) the order of the evaluation is not changed 2) for this to work,
no default value should be set in PROVISIONING_INTERFACE prior to the
execution of wait_for_interface_or_ip; thus remove "provisioning" from
get_provisioning_interface.

NOTE: a test for PROVISIONING_MACS could also be moved to
wait_for_interface_or_ip and get_provisioning_interface removed, but
this commits aims at minimizing change.

Signed-off-by: Marco Chiappero <marco.chiappero@suse.com>