Centralize assignment of HTTP_PORT
This is actually a bug fix, as ironic-common.sh, when invoked
without setting HTTP_PORT, will fail, as no default value is
defined.
We set HTTP_PORT default value in ironic-common.sh to 80 as
correctly written in the docs, and then export it.
Removing assignments from all other scripts to avoid mistakes
and confusion.

Signed-off-by: Riccardo Pittau <elfosardo@gmail.com>