Change default json-rpc port to avoid possible conflicts
The default ironic json-rpc port is 8089
In certain scenarios (e.g. HA) this port may conflict with other
services (e.g. k-nmstate).
Even if it's a rare occurrence, we should change the default
port in metal3 to avoid such conflicts.
This patch introduces a variable that allows changing the json-rpc
port in the ironic configuration with a default value of 6189.

Signed-off-by: Riccardo Pittau <elfosardo@gmail.com>