Introduce IRONIC_IPV6 to bind on IPv6 sockets
The ironic scripts can use PROVISIONING_IP or PROVISIONING_INTERFACE as
input to determine an IP address for process binding. This results in
IRONIC_IP being defined once such an IP address is determined, which can
hold either an IPv4 or an IPv6 address as value.

This design is likely rooted in the assumption is that on Linux
IPv4-mapped IPv6 addresses can be leveraged to serve both IPv4 and IPv6
through a single socket. However this is not a good practice and two
separate sockets should be used instead whenever possible. As a
consequence we should detect whether to bind on IPv4, IPv6 or both, and
define separate variables for the purpose.

This change modifies the IRONIC_IP value deduction logic:
- introducing the variable IRONIC_IPV6 alongside the existing IRONIC_IP
- leaving IRONIC_IP for IPv4 addresses only
- attempting to populate both variables whenever possible

At the end of the evaluation process, at least one variable between
IRONIC_IP and IRONIC_IPV6 is expected to contain a valid value.

While introducing this new logic, also add the new get_ip_of_interface()
utility function to streamline the retrieval of the IPv4 and IPv6
addresses for a given interface.

Signed-off-by: Marco Chiappero <marco.chiappero@suse.com>